[BUGFIX] Reset page tree object before fetching the page-tree for a page-id.
The getTree method of t3lib_pageTree class determines the page tree for a
given page-id and stores it in an internal array. Multiple invokation of
this method appends the new results in this internal array. As a result,
at any time, the object holds all the previously determined trees.

The expandExcludeString method of tx_crawler_lib uses this method to
expand the exclude PIDs. Due to the above mentioned problem, the tree for
expand PIDs contained results from the previously determined trees, more
specifically from the previously processed crawler configuration, thereby
excluding them from processing in the current configuration.

Thanks to Nikola Stojiljkovic for pointing out the bug.

Change-Id: I7b5a955656f06bfcf09bba9f0d8b290d1c615f26