[BUGFIX] improve performance of mark as delete query
The performance of the CLI is bad, if the queue and process are big.
The reason was, the the "mark as deleted" query of the process items
is very bad (build up cartesian product) and took up to 1 minute and
work on over 60 million rows.

The patch reduce the load on the db heavily. We add a "deleted=0" to
the query (there is no reason to set already deleted items to deleted)
and alter the process table to use a index.
As result the dependent subquery is not called for the deleted process
items anymore and the query run in a few ms

Resolves: #61502
Change-Id: I142812c2393742a836b8cea34ef11609b483040c
Reviewed-on: http://review.typo3.org/36652
Reviewed-by: Joh. Feustel <s@feustel.eu>
Reviewed-by: Tomas Norre Mikkelsen <tomasnorre@gmail.com>
Reviewed-by: Stefan Rotsch <stefan.rotsch@aoemedia.de>
Reviewed-by: Tim Lochmüller <tim@fruit-lab.de>
Tested-by: Tim Lochmüller <tim@fruit-lab.de>
Tested-by: Stefan Rotsch <stefan.rotsch@aoemedia.de>