[TASK] Adapt update-join statements
DBAL does not support joins for update statements (as of 2019-05-08),
therefore a statement like this wont do anything useful:

```
$queryBuilder
	->update('table', 't')
	->(join|leftJoin|rightJoin)(
		'table2', 't2', 't.id = t2.t_id')
	->where('t2.id = 2')
	->set('t.foo', 'bar')
	->execute();
```

This results in a database error, telling you that t2.id is not defined,
and you would see that the generated SQL does not contain the expected
join statement.

A workaround for this is to work with a string in the where condition,
like this:

```
//...
	->where(
		't.id IN(SELECT t2.t_id FROM table2 t2 WHERE t2.id = 2)'
	)
```

This is quite ugly but it's the best we can do for now as long as
Doctrine DBAL does not provide a proper join for update.