[BUGFIX] Composer: Move system dependency
The package requires the PHP extension »posix« to be installed
on the system. The posix extension however is not available for Windows.

This causes failed installations of crawler on Windows systems.

Only suggest the installation of the extension and dont make it
a requirement anymore to allow installations on both Unix and Windows.

Refs #384