[TASK] Rework backend module to use Fluid (#414)
The Backend module is rewritten using Fluid
and primitive action methods to render most
output on the Info Module => Crawler.

Base styling is available, but some minor changes
within the ProcessOverview need some improvements.

The old template functionality (incl pagination,
as this was not used anymore anyways) for ProcessOverview
is removed. Additional icons and css files are
removed as well, and TYPO3 Backend styling is used.

Usages of `global $LANG` is removed, most
rendering except for dropdowns is now built in fluid.

Rendering of nested columns within the Log view is still
compiled in the controller.

Controller does not inherit from AbstractFunctionModule
anymore as this class was deprecated in TYPO3 v9.

The tests are not needed anymore as there is no CLI user for crawler anymore.