[TASK] Add resname attribute to trans-unit tags in XLF files (#605)
Crowdin doesn't show the information stored in the attribute "id".
If it should be visible, the attribute must be duplicated using
"resname".

find Resources/Private/Language/ -name '*.xlf' -exec sed -i -e \
    '/resname/!s/\(<trans-unit id="\)\([^"]*\)"/\1\2" resname="\2"/' \
    '{}' +

Additionally, the attribute original is filled with the path of the translation
file instead of dummy value "messages". This string is also shown in Crowdin
and helps users while translating.