[BUGFIX] suppress unserialize E_WARNINGS
Fixes #1087

The intention of the try-catch block was to
catch Throwables during the unserialization of $dataString.

However, unserialize only throws exceptions when the
initialization of objects fails. With 'allowed_classes' => false,
no objects will be initialized.

Since PHP 8.3, an E_WARNING is issued if the string
is not unserializable. As we are not certain
if the string contains serialized data, we only want
to attempt to unserialize it.

To prevent E_WARNINGS, we need to suppress errors
instead of using try-catch.