[BUGFIX] Fix showing "Scheduled" and "Run-time" columns in backend log
In Info > Site Crawler > Crawler log:
As soon as the backend user activated the "Show Result Log" checkbox
one time, the columns "Scheduled" and "Run-time" are empty when
the checkbox is disabled.

The reason is that the variable's value is not taken into account,
only its existence in the user session data.