[BUGFIX] Fix indexing of cached pages when "disableFrontendIndexing" … (#1142)
* [BUGFIX] Fix indexing of cached pages when "disableFrontendIndexing" is active

When pages are cached in TYPO3v12, the indexing process is not
triggered when crawler fetches the pages.
This becomes a problem when the "Disable Indexing in Frontend"
(basic.disableFrontendIndexing) indexed_search flag is enabled,
because no indexing is happening then.

The solution is to tell TYPO3 to not use the cache when the page
is requested by the crawler process.

See the ShouldUseCachedPageDataIfAvailableEvent documentation at
https://docs.typo3.org/c/typo3/cms-core/main/en-us/Changelog/12.0/Feature-96968-PSR-14EventForAvoidLoadingFrontendPagesFromCache.html

Resolves: https://github.com/tomasnorre/crawler/issues/1029

* [BUGFIX] Remove unneeded initialization of queueRepository

phpStan complained:
> Property AOE\Crawler\Domain\Model\Process::$queueRepository on left
> side of ??= is not nullable nor uninitialized.
>  nullCoalesce.initializedProperty

The variable gets initialized in the constructor already.