[FEATURE] Extract error information in SubProcess crawl strategy (#1161) (#1162)
Up to now, the sub process execution strategy did not detect errors.
The HTML of error pages was provided as content, and the backend showed
an "OK" status.

Since cli/bootstrap.php does not set a non-zero exit code on errors,
we have to parse the HTML ourselves to detect if an error occured.

This patch supports TYPO3 error pages and extracts HTTP status code,
status message and error message from it.

Error information is logged into the debug log as well as
made visible in the backend site crawler log view.