[BUGFIX] Clear page cache when crawling for re-indexing (#1165)
When "basic.disableFrontendIndexing" is activated, TYPO3 does not index
crawled pages when they are already cached.

This was fixed for TYPO3v12+ in [1] by using the
"ShouldUseCachedPageDataIfAvailableEvent".
That event is not available in TYPO3v11.

This patch simply clears the frontend cache of a page
directly before crawling it.

[1]: https://github.com/tomasnorre/crawler/pull/1142
Related: https://github.com/tomasnorre/crawler/issues/1029